WebApp/requirements.txt
flask
flask-login
flask_sqlalchemy
sqlalchemy_imageattach
werkzeug.security
sqlalchemy
mysql-connector-python